feat(nft)：新增接口；修改绑定钱包的数据结构

+++ project/app/src/main/java/tech/sud/mgp/hello/ui/main/nft/widget/dialog/ChangeInternalAccountDialog.java
@@ -102,7 +102,7 @@ public class ChangeInternalAccountDialog extends BaseDialogFragment {
         }
     }
 
-    private boolean isUseWallet(int walletType) {
+    private boolean isUseWallet(long walletType) {
         BindWalletInfoModel bindWalletInfo = viewModel.getBindWalletInfo();
         if (bindWalletInfo != null) {
             return bindWalletInfo.walletType == walletType;
